Extract nested visitors; handle self assignments
Move nested SyntaxVisitor helpers (MemberVisitor, ReturnVisitor, BareMemberVisitor) out of SourceVisitor into a new SourceVisitor+NestedVisitors.swift file. Add assignmentTargetName(_:) and update assignment handling to recognize bare names and implicit-self (self.x) targets while ignoring member accesses on other objects. Add tests to verify self-qualified property assignments are tracked and foreign-object assignments are ignored.
